reorganize repo, make python write logfile
