--- rapid_process.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
--/usr/lib/tethys/bin/python /home/alan/work/scripts/spt_ecmwf_autorapid_process/run.py 1> /home/alan/work/logs/rapid_$(date +%y%m%d%H%M%S).log 2>&1
