+++ spt_ecmwf_autorapid_process/__init__.py
@@ -1,0 +1,1 @@
+from rapid_process import run_ecmwf_rapid_process
