+++ spt_ecmwf_autorapid_process/htcondor_ecmwf_rapid.py
