+++ spt_ecmwf_autorapid_process/setup/__init__.py
@@ -1,0 +1,1 @@
+from .create_cron import create_cron
